CORS and exposed header changes

+++ src/main/java/com/example/patients_portal/repository/PatientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PatientRepository extends CrudRepository<Patient, Long> {
+public interface PatientRepository extends CrudRepository<Patient, Integer> {
     Patient findById(int Id);
     Boolean deleteById(int Id);
     //List<Patient> findByIdAndName(int Id, String name);
